# Speed up queries for average recitation score.