--- a/test/models/recitation_section_test.rb
+++ b/test/models/recitation_section_test.rb
@@ -60,4 +60,46 @@ class RecitationSectionTest < ActiveSupport::TestCase
       @recitation.update! params
     end
   end
+
+  describe '#average_metric_score' do
+    let(:metric) { assignment_metrics(:assessment_overall) }
+    let(:gradeless_user) { users(:mandark) }
+
+    describe 'some students in the section have no grades' do
+      before do
+        gradeless_user.registration_for(courses(:main)).update!(
+            recitation_section: recitation)
+        assert_empty metric.grades.where(subject: gradeless_user)
+      end
+
+      it 'returns the average student score for the given metric only, excluding
+          students without grades' do
+        assert_equal 3, recitation.users.count
+        assert_equal 3, metric.grades.count
+        assert_equal 2, metric.grades.where(subject: recitation.users).count
+        assert_equal (70 + 10) / 2.0, recitation.average_metric_score(metric)
+      end
+    end
+
+    describe 'none of the students in the section have grades' do
+      before { metric.grades.where(subject: recitation.users).destroy_all }
+
+      it 'returns 0' do
+        assert_equal 0, recitation.average_metric_score(metric)
+      end
+    end
+
+    describe 'the section has no students' do
+      before do
+        recitation.users.each do |u|
+          u.registration_for(courses(:main)).update!(recitation_section: nil)
+        end
+        recitation.users.reload
+      end
+
+      it 'returns 0' do
+        assert_equal 0, recitation.average_metric_score(metric)
+      end
+    end
+  end
 end