Testing results on seed data:
* Sped up from 1034.6ms for **289** SQL queries, to 196.3ms for **32** SQL queries.
* The total page load went from **2189ms** down to **1267ms**